backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Settings management ───────────────────────────────────────────────────
pydantic-settings>=2.2.0  

# ── Document extractors ───────────────────────────────────────────────────
PyPDF2>=3.0.1
python-docx>=1.1.2
python-pptx>=0.6.23

# ── Image handling (required for Gemini image OCR) ────────────────────────
Pillow>=10.3.0

# ── Google Gemini ─────────────────────────────────────────────────────────
google-genai>=0.8.0

# ── Async HTTP client (direct Gemini REST calls, bypasses SDK routing) ────
httpx>=0.27.0

# ── File upload support ───────────────────────────────────────────────────
python-multipart>=0.0.9

# ── Supabase ──────────────────────────────────────────────────────────────
# The backend uses native requests calls directly against Supabase's REST/Storage
# API — the supabase-py SDK is NOT needed and was causing httpx version conflicts.
requests>=2.31.0
